move some utils function from auto-consensus to auto-utils
