+++ packages/auto-consensus/src/types/events.ts
@@ -1,7 +1,6 @@
 // file: src/types/events.ts
 
 export type ActionEvents = string | string[]
-export type Events = ActionEvents | ActionEvents[]
 
 // Define specific extrinsic keys for events
 export type EventKeys =
