+++ packages/auto-consensus/src/types/index.ts
@@ -1,4 +1,3 @@
 // file: src/types/index.ts
 
 export * from './events'
-export * from './transaction'
