+++ packages/auto-consensus/src/utils/index.ts
@@ -1,7 +1,4 @@
 // file: src/utils/index.ts
 
-export * from './detectTxSuccess'
 export * from './events'
-export * from './signAndSendTx'
 export * from './sudo'
-export * from './validateEvents'
