+++ packages/auto-utils/src/types/event.ts
@@ -1,3 +1,6 @@
 import type { EventRecord } from '@polkadot/types/interfaces'
 
+export type ActionEvents = string | string[]
+export type Events = ActionEvents | ActionEvents[]
+
 export type { EventRecord }
