+++ packages/auto-utils/src/types/index.ts
@@ -3,6 +3,7 @@ import type { BN } from '@polkadot/util'
 export * from './event'
 export * from './extrinsic'
 export * from './network'
+export * from './transaction'
 export * from './wallet'
 
 export type { BN }
