+++ packages/auto-utils/src/utils/index.ts
@@ -1,5 +1,8 @@
 export * from './createType'
+export * from './events'
 export * from './format'
 export * from './ready'
 export * from './sign'
+export * from './signAndSendTx'
+export * from './validateEvents'
 export * from './verify'
