+++ packages/auto-utils/src/utils/validateEvents.ts
@@ -1,8 +1,6 @@
 // file: src/utils/validateEvents.ts
 
-import type { EventRecord } from '@autonomys/auto-utils'
-import type { Events } from '../types/events'
-import type { EventsValidated } from '../types/transaction'
+import type { EventRecord, Events, EventsValidated } from '../types'
 import { expectSuccessfulTxEvent } from './events'
 
 export const validateEvents = (
